% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest-package}
\alias{manifest-package}
\title{manifest: Data Provenance Tracking for Research Databases}
\description{
Comment conventions:
\verb{#:}  Manifest criteria  --- captured as CRITERIA in the log
\verb{##}  Internal/dev notes --- ignored by manifest
\verb{#}   Ordinary comments  --- ignored by manifest
}
