% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_log}
\alias{manifest_log}
\title{Inspect the current manifest log mid-script}
\usage{
manifest_log()
}
\value{
Invisibly returns the log data frame.
}
\description{
Inspect the current manifest log mid-script
}
